Fix PR file-size gate after main sync
