+++ .pm/tasks/task_acb7e3599b4242628a7ac99a62628d55.execution.md
@@ -165,3 +165,12 @@ Example:
 - Expected Result: Terminology scan passes for only current compatibility content/path hits; temporary legacy-named probe fails; shell syntax passes; Rust file-size gate reproduces locally if task diff caused the CI failure; branch diff does not touch the CI-reported oversized Rust files; doc and PR-ready workflow gates pass.
 - Actual Result: Terminology scan exited `0` with `allowed compatibility content hits: 17` and `allowed compatibility path hits: 17`; temporary legacy-named probe exited `1` and reported `legacy terminology in path name`; `bash -n` exited `0`; local `check-rust-file-size.sh` exited `0`; diff probe exited `1` with no matching changed files; `doc-governance-check: OK`; `workflow-lint: OK (task_acb7e3599b4242628a7ac99a62628d55, phase=pr-ready)`.
 - Blocker / Next Action: Run task-local workflow/doc/diff verification, commit and push the PR feedback fix, then continue PR watch/fix/merge loop.
+
+## 2026-06-17 12:44:00 CST / tpm
+- 完成内容: PR #500 Rust required-gate failure reproduced and fixed after syncing current `origin/main` into the task branch. GitHub required-gate used the PR merge ref and failed `./scripts/check-rust-file-size.sh` on files introduced by the newer `main` tip after this branch forked: `letai_direct.rs`, `runtime_live.rs`, `oasis7_game_launcher_tests.rs`, `runtime_live/tests.rs`, and `runtime_live/tests/auth_actions.rs`. Because required checks failed, TPM continued the normal PR watch/fix loop and dispatched bounded post-PR slices.
+- 遗留事项: Commit and push this CI-fix merge/remediation, then re-watch GitHub Rust/Wasm checks, PR comments, review threads, and mergeability. Residual local risk: Cargo verification was slow due shared build/package locks and concurrent repo-family builds, but all targeted list/compile probes eventually passed.
+- Action: Merged `origin/main` into `task/runtime-unified-world-code-layer-migration-follow-up`; integrated `repository_health_engineer` slice `019ed390-be3a-7192-8525-ae50111f738f` for provider/launcher file-size remediation and `viewer_engineer`/`repository_health_engineer` slice `019ed390-f49d-7032-9361-6af8fd0d8180` for runtime-live remediation. Follow-up TPM fixes corrected extracted-module visibility/import boundaries found by Cargo: `std::io` import in `runtime_live/support.rs`, LetAI topup helper visibility/re-export shape, and `TestChainStatusServer` re-export visibility.
+- Validation Command: `./scripts/check-rust-file-size.sh`; `cargo fmt --check`; `git diff --check`; `./scripts/cargo-dev.sh test -p oasis7 --bin oasis7_game_launcher -- --list`; `./scripts/cargo-dev.sh test -p oasis7 --bin oasis7_provider_local_bridge -- --list`; `./scripts/cargo-dev.sh test -p oasis7 runtime_live -- --list`; line-count probe for the five CI-reported oversized files.
+- Expected Result: Rust file-size gate reports zero oversized/structural files; formatting and diff whitespace pass; affected launcher/provider/runtime-live modules compile and list their tests; previously oversized files fall below 1200 lines.
+- Actual Result: `check-rust-file-size: OK` with `oversized code files=0, test files=0, structural slice files=0, include targets=0`; `cargo fmt --check` exited `0`; `git diff --check` exited `0`; launcher discovery exited `0` and listed 110 tests including `launcher_static_http_tests::static_http_server_serves_large_static_asset_completely`; provider bridge discovery exited `0` and listed 47 tests including LetAI auto-topup tests; runtime-live discovery exited `0` and listed 115 tests including extracted `chain_sync::*` and `auth_actions::agent_chat_tests::*`. Final line counts: `letai_direct.rs` 989, `oasis7_game_launcher_tests.rs` 1165, `runtime_live.rs` 889, `runtime_live/tests.rs` 935, `runtime_live/tests/auth_actions.rs` 893.
+- Blocker / Next Action: Run final post-PR workflow/doc/diff checks, commit and push, then continue PR #500 CI/comment/merge watch.
