+++ crates/oasis7/src/bin/oasis7_game_launcher/oasis7_game_launcher_tests.rs
@@ -1,7 +1,7 @@
 use std::env;
 use std::fs;
-use std::io::{BufRead, BufReader, BufWriter, Read, Write};
-use std::net::{TcpListener, TcpStream};
+use std::io::{BufRead, BufReader, BufWriter, Write};
+use std::net::TcpListener;
 use std::path::Path;
 use std::process::Command;
 use std::thread;
@@ -39,6 +39,8 @@ use oasis7_proto::storage_profile::StorageProfile;
 #[path = "viewer_static_dir_tests.rs"]
 mod viewer_static_dir_tests;
 use viewer_static_dir_tests::make_temp_dir;
+#[path = "launcher_static_http_tests.rs"]
+mod launcher_static_http_tests;
 
 fn assert_removed_old_brand_viewer_auth_env_absent(text: &str) {
     assert!(!text.contains(removed_old_brand_viewer_auth_bootstrap_object().as_str()));
@@ -1086,50 +1088,6 @@ fn sanitize_index_html_for_embedded_server_injects_viewer_auth_bootstrap_into_no
     assert!(sanitized.contains("viewer-player"));
     assert!(sanitized.contains("pub-hex"));
     assert!(sanitized.contains("priv-hex"));
-}
-
-#[test]
-fn static_http_server_serves_large_static_asset_completely() {
-    let temp_dir = make_temp_dir("large_static_asset");
-    let large_body = vec![b'a'; 512 * 1024];
-    fs::write(temp_dir.join("viewer.js"), &large_body).expect("write large asset");
-
-    let probe = TcpListener::bind(("127.0.0.1", 0)).expect("bind port probe");
-    let port = probe.local_addr().expect("probe addr").port();
-    drop(probe);
-
-    let mut server = start_static_http_server(
-        DeploymentMode::TrustedLocalOnly,
-        "127.0.0.1:0",
-        "127.0.0.1",
-        port,
-        temp_dir.as_path(),
-        None,
-    )
-    .expect("start static HTTP server");
-
-    let mut response = Vec::new();
-    for _ in 0..50 {
-        match TcpStream::connect(("127.0.0.1", port)) {
-            Ok(mut stream) => {
-                stream
-                    .write_all(b"GET /viewer.js HTTP/1.1\r\nHost: 127.0.0.1\r\n\r\n")
-                    .expect("write request");
-                stream.read_to_end(&mut response).expect("read response");
-                break;
-            }
-            Err(_) => thread::sleep(std::time::Duration::from_millis(20)),
-        }
-    }
-    stop_static_http_server(&mut server);
-
-    let split_at = response
-        .windows(4)
-        .position(|window| window == b"\r\n\r\n")
-        .expect("response headers end")
-        + 4;
-    assert!(String::from_utf8_lossy(&response[..split_at]).starts_with("HTTP/1.1 200 OK"));
-    assert_eq!(&response[split_at..], large_body.as_slice());
 }
 
 #[test]
