+++ crates/oasis7/src/viewer/runtime_live/support.rs
@@ -1,5 +1,5 @@
 use std::collections::HashSet;
-use std::io::Write;
+use std::io::{self, Write};
 
 use super::*;
